+++ Realm/Realm/Sync/ErrorHandling/DiscardLocalResetHandler.cs
@@ -20,12 +20,13 @@ namespace Realms.Sync.ErrorHandling
 {
     /// <summary>
     /// A client reset strategy where all the not yet synchronized data is automatically discarded and a fresh copy of the synchronized Realm is obtained.
-    /// To be noted that the freshly downloaded copy of the synchronized Realm triggers all change notifications as a write transaction is internally simulated.
-    /// This strategy supplies three callbacks: <see cref="OnBeforeReset"/> and <see cref="OnAfterReset"/> to give a chance to take some user actions,
-    /// respectively, before and after a client reset has happened.
-    /// And in case something goes wrong during the client reset, <see cref="ManualResetFallback"/> is supplied to take specific actions in such condition.
-    /// For more information see <see cref="ClientResetHandlerBase"/>.
     /// </summary>
+    /// <remarks>
+    /// The freshly downloaded copy of the synchronized Realm triggers all change notifications as a write transaction is internally simulated.
+    /// This strategy supplies three callbacks: <see cref="OnBeforeReset"/>, <see cref="OnAfterReset"/>, and <see cref="ManualResetFallback"/>.
+    /// The first two give notify you just before and after the client reset has happened,
+    /// while the last one will be invoked in case an error occurs during the automated process and the system needs to fallback to a manual mode.
+    /// </remarks>
     /// <seealso href="https://docs.mongodb.com/realm/sync/overview/">Sync Overview Docs</seealso>
     public sealed class DiscardLocalResetHandler : ClientResetHandlerBase
     {
@@ -44,24 +45,27 @@ namespace Realms.Sync.ErrorHandling
         /// The frozen <see cref="Realm"/> as it was before the reset.
         /// </param>
         /// <param name="after">
-        /// The <see cref="Realm"/> after the client reset.
+        /// The <see cref="Realm"/> after the client reset. In order to modify this realm a write transaction needs to be started.
         /// </param>
         public delegate void AfterResetCallback(Realm beforeFrozen, Realm after);
 
         /// <summary>
         /// Gets or sets the callback that indicates a Client Reset is about to happen.
-        /// Among other things, you can use this call to temporarily store the before Realm as a backup and in the <see cref="OnAfterReset"/> callback merge the changes, if necessary.
         /// </summary>
+        /// <value>Callback invoked right before a Client Reset.</value>
         public BeforeResetCallback OnBeforeReset { get; set; }
 
         /// <summary>
-        /// Gets or sets the callback that indicates a Client Reset just happened. Special custom actions can be taken at this point like merging local changes if the "before" realm was stored during <see cref="BeforeResetCallback"/>.
+        /// Gets or sets the callback that indicates a Client Reset just happened.
+        /// Special custom actions can be taken at this point like merging local changes from <value>beforeFrozen</value>.
         /// </summary>
+        /// <value>Callback invoked right after a Client Reset.</value>
         public AfterResetCallback OnAfterReset { get; set; }
 
         /// <summary>
         /// Gets or sets the callback triggered when an error has occurred that makes the operation unable to complete, for example in the case of a destructive schema change.
         /// </summary>
+        /// <value>Callback invoked if a Client Reset fails.</value>
         public ClientResetCallback ManualResetFallback { get; set; }
     }
 }