+++ Tests/Realm.Tests/Sync/SyncTestHelpers.cs
@@ -84,7 +84,6 @@ namespace Realms.Tests.Sync
             {
                 var tcs = new TaskCompletionSource<object>();
 
-#pragma warning disable CS0618 // Type or member is obsolete
                 Session.Error += HandleSessionError;
                 try
                 {
@@ -94,7 +93,6 @@ namespace Realms.Tests.Sync
                 {
                     Session.Error -= HandleSessionError;
                 }
-#pragma warning restore CS0618 // Type or member is obsolete
 
                 void HandleSessionError(object _, ErrorEventArgs errorArgs)
                 {
@@ -313,15 +311,10 @@ namespace Realms.Tests.Sync
                     tcs.TrySetException(ex);
                 }
 
-#pragma warning disable CS0618 // Type or member is obsolete
                 Session.Error -= handler;
-#pragma warning restore CS0618 // Type or member is obsolete
             });
 
-#pragma warning disable CS0618 // Type or member is obsolete
             Session.Error += handler;
-#pragma warning restore CS0618 // Type or member is obsolete
-
             return tcs.Task;
         }
     }