+++ CHANGELOG.md
@@ -1,7 +1,41 @@
 ## vNext (TBD)
 
 ### Enhancements
-* Added support for a new client reset strategy, called ["Discarding local"](add the right link to docs here). Such new stragegy greatly simplifies the whole handling of a client reset on a synchronized Realm.
+* Added support for a new client reset strategy, called [Discard Local Reset](www.mongodb.com/FIXME_PLEASE). This new stragegy greatly simplifies the handling of a client reset event on a synchronized Realm.  
+This addition makes `Session.Error` **deprecated**. In order to migrate the existing client reset handling code to the new structure the following should be done
+  ```csharp
+    var conf = new PartitionSyncConfiguration(partition, user)
+    {
+      ClientResetHandler = new ManualRecoveryHandler
+      {
+          OnClientReset = (sender, e) =>
+          {
+              // user's code that's being migrated
+          }
+      }
+    };
+  ```
+  And if, instead, one wants to have the handling code to be used only if a discarding local changes fails, the following should be done instead
+  ```csharp
+    var conf = new PartitionSyncConfiguration(partition, user)
+    {
+      ClientResetHandler = new DiscardLocalResetHandler
+      {
+          OnBeforeReset = (beforeFrozen) =>
+          {
+            // executed right before a client reset is about to happen
+          },
+          OnAfterReset = (beforeFrozen, after) =>
+          {
+            // executed right after a client reset is has completed
+          },
+          ManualResetFallback = (session, err) =>
+          {
+              // user's code that's being migrated
+          }
+      }
+    };
+  ```
 * Lifted a limitation that would prevent you from changing the primary key of objects during a migration. It is now possible to do it with both the dynamic and the strongly-typed API:
   ```csharp
   var config = new RealmConfiguration